# server/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "aiscast API",
    "version": "1.0.0",
    "description": "Live AIS data from openwaters.io: current vessel positions, receiving stations, usage stats, self-serve tokens, and HTTP ingest for feeders. Successful responses are JSON; errors and `/health` are plain text. CORS is open, times are UTC, and JSON is compressed (zstd or gzip) when the request sends `Accept-Encoding`.\n\nThe streams are documented separately: `/v0/stream` (WebSocket, aisstream.io-compatible), `/v1/stream` (WebSocket JSON frames, MQTT over WebSocket for feeders, or Server-Sent Events), `/v1/nmea` (WebSocket, raw NMEA), and raw NMEA ingest on UDP port 10110. `/mcp` is the MCP (Model Context Protocol) endpoint for AI assistants: Streamable HTTP, no sign-in, the same tiers as `/v1/vessels`. See [openwaters.io/api/ais](https://openwaters.io/api/ais/).",
    "contact": {
      "name": "aiscast",
      "url": "https://github.com/openwatersio/aiscast"
    },
    "termsOfService": "https://github.com/openwatersio/aiscast/blob/main/docs/contributor-agreement.md",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    }
  },
  "servers": [{ "url": "https://ais.openwaters.io" }],
...
  "paths": {
...
    "/v1/receive": {
      "post": {
        "operationId": "receive",
        "tags": ["Contributing"],
        "summary": "Contributing over HTTP",
        "description": "Contributing over HTTP with AIS-catcher's `-H` output or your own client:\n\n```\nAIS-catcher -H https://ais.openwaters.io/v1/receive USERPWD x:<token> GZIP on INTERVAL 15\n```\n\nThe body is AIS-catcher's `jsonaiscatcher` envelope or plain newline-separated NMEA, optionally gzipped, up to 1 MB. Needs a token, whose `sub` becomes the station name. 600 posts per minute per station. Feeding data is acceptance of the contributor agreement, linked from every response's `Link: rel=\"terms-of-service\"` header.",
        "security": [{ "token": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "description": "AIS-catcher `jsonaiscatcher` envelope.",
                "properties": {
                  "msgs": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "nmea": { "type": "array", "items": { "type": "string" } },
                        "rxtime": { "type": "string", "description": "YYYYMMDDHHMMSS UTC; used as the message time when within 30 s of arrival." },
                        "channel": { "type": "string" }
                      }
                    }
                  }
                }
              },
              "example": { "msgs": [{ "nmea": ["!AIVDM,1,1,,A,13HOI:0P0000VOHLCnHQKwvL05Ip,0*23"], "rxtime": "20260820111900", "channel": "A" }] }
            },
            "text/plain": {
              "schema": { "type": "string", "description": "Newline-separated NMEA sentences, TAG blocks allowed." },
              "example": "!AIVDM,1,1,,A,13HOI:0P0000VOHLCnHQKwvL05Ip,0*23"
            }
          }
        },
        "responses": {
          "200": { "description": "Accepted. The body is empty." },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "413": {
            "description": "Body larger than 1 MB (before or after gzip decompression).",
            "content": { "text/plain": { "schema": { "type": "string" } } }
          },
          "429": { "$ref": "#/components/responses/RateLimited" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "token": {
        "type": "http",
        "scheme": "bearer",
        "description": "An aiscast claim token (`ak1....`), created by `POST /v1/keys` or issued by the operator. Besides `Authorization: Bearer`, the token is accepted as an HTTP Basic password (`anything:<token>`, which is what AIS-catcher's USERPWD sends) or as `?key=<token>` on the URL. A token that is present but invalid, expired, revoked, used outside its `cidr`, or lacking the role for the action is refused; it never degrades to anonymous access."
      }
    },
...
    "responses": {
      "BadRequest": {
        "description": "Malformed parameter or body. The plain-text body says which.",
        "content": { "text/plain": { "schema": { "type": "string" } } }
      },
      "Unauthorized": {
        "description": "Missing, invalid, expired, or revoked token, or the token lacks the role for this action. The plain-text body gives the reason.",
        "content": { "text/plain": { "schema": { "type": "string" } } }
      },
      "RateLimited": {
        "description": "Over the per-address or per-station request limit.",
        "content": { "text/plain": { "schema": { "type": "string" }, "example": "rate limited" } }
      }
    },
...
  }
}